refactor: move file-manipulation to own module
