+++ src/MHTimer.js
@@ -1,6 +1,6 @@
-/*
-  MHTimer Bot
-*/
+/**
+ * MHTimer Bot
+ */
 // Import required modules
 const { DateTime, Duration, Interval } = require('luxon');
 const Discord = require('discord.js');
@@ -16,21 +16,15 @@ const {
     prettyPrintArrayAsString,
     splitString,
     timeLeft,
-} = require('./modules/utils');
+} = require('./modules/format-utils');
+const { loadDataFromJSON, saveDataAsJSON } = require('./modules/file-utils');
 const Logger = require('./modules/logger');
 
-// Access local URIs, like files.
-const fs = require('fs');
 // Access external URIs, like @devjacksmith 's tools.
 const fetch = require('node-fetch');
 const { URLSearchParams } = require('url');
 // We need more robust CSV handling
 const csv_parse = require('csv-parse');
-
-// Convert callbacks to 'Promise' versions
-const util = require('util');
-const fs_readFile = util.promisify(fs.readFile);
-const fs_writeFile = util.promisify(fs.writeFile);
 
 // Globals
 const client = new Discord.Client({ disabledEvents: ['TYPING_START'] });
@@ -40,7 +34,6 @@ const main_settings_filename = 'data/settings.json',
     hunter_ids_filename = 'data/hunters.json',
     reminder_filename = 'data/reminders.json',
     nickname_urls_filename = 'data/nicknames.json';
-const file_encoding = 'utf8';
 
 const settings = {},
     mice = [],
@@ -303,43 +296,6 @@ function quit() {
         .catch(err => {
             Logger.error('Shutdown: unhandled error:\n', err, '\nImmediately exiting.');
             process.exit();
-        });
-}
-
-/**
- * Generic Promise-based file read.
- * Returns the data as an object, or the error that occurred when reading and parsing the file.
- * A common error code will be 'ENOENT' (the file did not exist).
- *
- * @param {string} filename the name of a file in the current working directory (or a path and the name)
- *                          from which raw data will be read, and then parsed as JSON.
- * @returns {Promise <any>}  Data from the given file, as an object to be consumed by the caller.
- */
-function loadDataFromJSON(filename) {
-    return fs_readFile(filename, { encoding: file_encoding })
-        .then(data => {
-            Logger.log(`I/O: data read from '${filename}'.`);
-            return data;
-        }).then(rawData => JSON.parse(rawData));
-}
-/**
- * Generic Promise-based file write.
- * Returns true if the file was written without error.
- * Returns false if an error occurred. Depending on the error, the data may have been written anyway.
- *
- * @param {string} filename the name of a file in the current working directory (or a path and the name)
- *                          to which data will be serialized as JSON.
- * @param {any} rawData raw object data which can be serialized as JSON, via JSON.stringify()
- * @returns {Promise <boolean>} The result of the save request (false negatives possible).
- */
-function saveDataAsJSON(filename, rawData) {
-    return fs_writeFile(filename, JSON.stringify(rawData, null, 1), { encoding: file_encoding })
-        .then(() => {
-            Logger.log(`I/O: data written to '${filename}'.`);
-            return true;
-        }).catch(err => {
-            Logger.error(`I/O: error writing to '${filename}':\n`, err);
-            return false;
         });
 }
